Move TriggerChangedEventArgs class to new file
